[88] Main 비지니스로직 재구성
